Create PuttingThingsTogether.py
This is a Combo of the Geopy lib def and the API def.
There is no GUI element yet, and it does not show the closest stop. need to add to the "get_nearby_bus_stops" function a set of code that parses the .json out, and looks at the outputs:
    "distance": 45,
    "global_stop_id": "1:94380",
    "location_type": 0,
    "parent_station_global_stop_id": "string",
    "route_type": 0,
    "stop_lat": 45.526168077787894,
    "stop_lon": -73.59506067289408,
    "stop_name": "De Gaspé / Maguire",
    "stop_code": "51752",
    "rt_stop_id": "51752",
    "wheelchair_boarding": 1
And shows us the "stop_lat," "stop_lon," and "stop_name" of the stop with the closest "distance" from the place the user input.