Update DisplaysRoutesAndDeparturesAtStop.py
ChangeLog:
added -4 hr to display EST.
Changed 24 to 12 hr time standard.
moved date to its own variable.
switched around exported time display.
Added API reader to keep the API off future code and to it's on txt file. In the code, if we want to call the API, just reference the text in the file: "API.txt" (this is so we do not accidentally post the API ky public).