Create main.py
Prints are now variables for GUI reference. 
Sorted departures and gave them variables to be referenced for the GUI. (code for Next departure, next 3, next 5, and all).
notes for the possibility of a count down timer to the next departure.

added back debug prints to see how it works along the way.